Finished the version that uses an optimizer to improve the candidate before sending to dReal.
Found that reducing the number of decimal places to which coefficients are computed helps the
process terminate much more quickly. I conjecture that this is because it can kill off coefficients
that should be zero much more quickly.

This may have side effects for some examples? At any rate, the user may specify the desired number
of decimal places, so this is not a big deal.

I also decided to exclude a smaller region around the origin than what I was excluding before. This
should improve the quality of the counterexamples that dReal sends back.